+++ test/doc_test.exs
@@ -2,5 +2,6 @@ defmodule Doc.Test do
   use ExUnit.Case
   doctest Cldr.List
   doctest MyApp.Cldr.List
+  doctest Cldr.List.Pattern
 
 end